requirements.txt

matplotlib
pyserial
